Added a check for runs directory existing

+++ driver/run_size.csh
@@ -6,7 +6,14 @@ setenv rundir $PWD
 setenv root $rundir/cytosim
 
 cd $rundir
-rm -rf runs/*
+if (! -d $rundir/runs) then
+  echo "runs directory does not exist"
+  mkdir runs;
+else
+  echo "runs directory exists, removing old run results"
+  rm -rf runs/*;
+endif
+ 
 $root/python/run/preconfig.py runs/run%04i/config.cym cfg/config.cym.tpl
 $root/python/look/scan.py '../../cytosim/bin/sim' runs/run????
 $root/python/look/scan.py '../../cytosim/bin/play image image_format=ppm frame=10' runs/run????
